pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-ecg-image-to-signal-reconstruction-pipeline"
version = "0.1.0"
description = "A maintainable ECG image digitization runtime with explicit model selection and preserved provenance."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Project Author" }]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
runtime = [
  "opencv-python-headless>=4.10",
  "pyarrow>=16.0",
  "scikit-image>=0.24",
  "scipy>=1.13",
  "segmentation-models-pytorch>=0.3.4",
  "torch>=2.2",
  "tqdm>=4.66",
  "ultralytics>=8.2",
  "matplotlib>=3.9",
]
dev = [
  "pytest>=8.2",
  "ruff>=0.6",
]

[project.scripts]
ecg-digitizer = "ecg_digitizer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["archive"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
